+++ packages/e2e-test/.eslintrc.js
@@ -1,6 +1,5 @@
 module.exports = {
   extends: [require.resolve('@backstage/cli/config/eslint.backend')],
-  ignorePatterns: ['templates/**'],
   rules: {
     'no-console': 0,
     'import/no-extraneous-dependencies': [