+++ plugins/scaffolder-backend/src/service/router.ts
@@ -136,7 +136,10 @@ export async function createRouter(
         },
       };
 
-      const template = await entityClient.findTemplate(templateName);
+      // Forward authorization from client
+      const template = await entityClient.findTemplate(templateName, {
+        token: getBearerToken(req.headers.authorization),
+      });
 
       const validationResult: ValidatorResult = validate(
         values,
@@ -307,4 +310,8 @@ export async function createRouter(
   app.use('/', router);
 
   return app;
+}
+
+function getBearerToken(header?: string): string | undefined {
+  return header?.match(/Bearer\s+(\S+)/i)?.[1];
 }