+++ packages/catalog-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/catalog-client
 
+## 0.3.6
+
+### Patch Changes
+
+- 6ed2b47d6: Include Backstage identity token in requests to backend plugins.
+- 72b96e880: Add the `presence` argument to the `CatalogApi` to be able to register optional locations.
+
 ## 0.3.5
 
 ### Patch Changes